+++ integrations/adobe-analytics/HISTORY.md
@@ -1,3 +1,12 @@
+1.15.2 / 2020-02-12
+===================
+
+  * Support sending custom Context Data Variables on `Video Playback Started`, `Video Playback Resumed`, and `Video Content Started` when using Adobe Heartbeat Tracking URL (from context object)
+
+1.15.1 / 2020-02-12
+===================
+
+  * Support sending custom Context Data Variables on `Video Playback Started`, `Video Playback Resumed`, and `Video Content Started` when using Adobe Heartbeat Tracking URL (from properties object)
 
 1.14.3 / 2018-06-25
 ===================